confirm action delete medic

+++ frontend-care-tym/src/Component/Center/Medic/ListEditMedic.js
@@ -1,12 +1,13 @@
 import { useEffect, useState } from "react"
 import { useNavigate } from "react-router-dom"
 import { deleteRequest, getRequest } from "../../../Services/Request"
+import Swal from 'sweetalert2'
 
 export default function ListEditMedic(props) {
 
     const [data, setData] = useState([])
     const [loading, setLoading] = useState(true)
-    
+
     const api = props.api
 
 
@@ -18,22 +19,53 @@ export default function ListEditMedic(props) {
         setLoading(false)
     }, [api])
 
-    const handleEdit = (x)=>{
-        props.setEdit(x) 
+    const handleEdit = (x) => {
+        props.setEdit(x)
         props.setType(2)
     }
 
-    const handleDelete = async(data,x) =>{
-        console.log()
-        const resp = await deleteRequest("http://127.0.0.1:8000/api/center/medic/"+ x.rut)
-        if (resp.status === 200){
-            window.location.reload()
-            //const deleteData = data.filter(data => data.rut !== x.rut)
-            //console.log(deleteData)
-            //setData(deleteData)
-
+    const handleDelete = async (data, x) => {
+        
+        const deleteAccion = await Swal.fire({
+            title: "Accion permanente",
+            text: "Eliminar medico con rut: " + x.rut,
+            icon: "warning",
+            showCancelButton: true,
+            cancelButtonText: "Cancelar",
+            cancelButtonColor: "#1D4ED8",
+            confirmButtonText: "Eliminar medico",
+            confirmButtonColor: "#D33"
+        })
+        if (deleteAccion.isConfirmed) {
+            const resp = await deleteRequest("http://127.0.0.1:8000/api/center/medic/" + x.rut)
+            if (resp.status === 200){
+                Swal.fire({
+                    title: "Medico con rut: " + x.rut + " eliminado con exito!",
+                    icon: "success"
+                })
+            }else{
+                Swal.fire({
+                    title: "Error desconocido, se recomienda actualizar la pagina.",
+                    icon: "error",
+                })
+            }
+        } else {
+            Swal.fire({
+                title: "Accion cancelada",
+                icon: "error",
+                confirmButtonColor: "#1D4ED8"
+            })
         }
+
+
+        //if (resp.status === 200){
+        //window.location.reload()
+        //const deleteData = data.filter(data => data.rut !== x.rut)
+        //console.log(deleteData)
+        //setData(deleteData)
+
     }
+
 
 
     return (
@@ -51,8 +83,8 @@ export default function ListEditMedic(props) {
                             <h1 className="font-normal">Telefono: {x.telefono}</h1>
                             <h1 className="font-normal">Duracion de atencion: {x.duracion_atencion}</h1>
                             <div>
-                                <button className="m-2 ml-0" onClick={()=>handleEdit(x)} >Editar</button>
-                                <button className="m-2" onClick={()=> handleDelete(data,x)}>Eliminar</button>
+                                <button className="m-2 p-1 ml-0 text-white border-2 border-blue-700 bg-blue-700 rounded-md font-medium" onClick={() => handleEdit(x)} >Editar</button>
+                                <button className="m-2 border-2 border-red-600 text-red-600 rounded-md p-1 font-medium" onClick={() => handleDelete(data, x)}>Eliminar</button>
                             </div>
                         </div>
                     ))
